requirements.test.txt
# Requirements for testing

pyyaml
djangorestframework~=3.0

pytest
pytest-pythonpath
pytest-cov
pytest-django
